Add videoProvider protocol
